refactor: reorganize chart templates directory
