+++ pkg/chart/templates/chart.qtpl
